collector_alpaca_stocks/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
websockets==13.1
pydantic==2.9.2
google-cloud-pubsub==2.23.0
python-dotenv==1.0.1
